LUN-3: Move Interfaces from exercise.ts
